feat : 응답 상세 조회 api 구현
