+++ common/src/main/java/com/exit/common/response/success/QuestionSuccessCode.java
@@ -25,6 +25,7 @@ public enum QuestionSuccessCode implements SuccessCode {
     COMMENT_GET_SUCCESS("QUESTION_OK_016", HttpStatus.OK , "댓글 조회 성공"),
     GET_POPULAR_POST("QUESTION_OK_017", HttpStatus.OK , "인기 게시글 조회 성공"),
     GET_MY_QUESTION_SUCCESS("QUESTION_OK_018", HttpStatus.OK, "내 질문 목록 조회 성공"),
+    GET_DETAIL_RESPONSE("QUESTION_OK_019", HttpStatus.OK, "응답 목록 조회 성공"),
     ;
 
     private final String code;
