+++ gateway-service/src/main/java/com/exit/gateway/controller/question/dto/response/question/QuestionDetailResponseDto.java
@@ -2,24 +2,14 @@ package com.exit.gateway.controller.question.dto.response.question;
 
 import com.exit.common.grpc.QuestionDetailResponse;
 
-import java.util.List;
-
 public record QuestionDetailResponseDto(
-        QuestionCreateResponseDto question,
-        List<ResponseDetailDto> responses,
-        Boolean hasNext
+        QuestionCreateResponseDto question
 ) {
     public static QuestionDetailResponseDto from(QuestionDetailResponse grpcResponse) {
         QuestionCreateResponseDto questionDto = QuestionCreateResponseDto.from(grpcResponse.getQuestion());
 
-        List<ResponseDetailDto> responseDtos = grpcResponse.getResponsesList().stream()
-                .map(ResponseDetailDto::from)
-                .toList();
-
         return new QuestionDetailResponseDto(
-                questionDto,
-                responseDtos,
-                grpcResponse.getHasNext()
+                questionDto
         );
     }
 }
